Move viz controls to viz panel
