+++ packages/webapp/src/pages/MainPage/components/ThemeViewer/ThemeViewer.base.tsx
@@ -11,6 +11,7 @@ import { useMemo } from 'react'
 import type { Graph } from '../../../../types.js'
 import { D3Chart } from './charts/D3/index.js'
 import { charts, VegaChart } from './charts/Vega/index.js'
+import { ControlPanel } from './ControlPanel/ControlPanel.js'
 import { FluentControls } from './FluentControls/index.js'
 import { D3Graph } from './graphs/D3/index.js'
 
@@ -65,6 +66,7 @@ export const ThemeViewer: FC<ThemeViewerProps> = ({
 		<div className="theme-wrapper">
 			<Pivot>
 				<PivotItem headerText="Example visualizations">
+					<ControlPanel />
 					<div className="example-grid">
 						{graphs.map((graph, i) => (
 							<div className="example-grid-item" key={`example-graph-${i}`}>
